ai-chatbot/requirements.txt
# AI Chatbot 依赖
Flask==2.3.3
langgraph==0.0.30
langchain==0.1.0
langchain-core==0.1.0
openai==1.12.0
anthropic==0.25.4
qianfan==0.3.3
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
sqlite3
# 注意：minimax 可能需要额外的包，这里使用 openai 兼容接口